offscreen-canvas/the-offscreen-canvas/initial.colour.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>OffscreenCanvas test: initial.colour</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>

<h1>initial.colour</h1>
<p class="desc">Initial state is transparent black</p>

<p class="notes">Output should be transparent black (not transparent anything-else), but manual
verification can only confirm that it's transparent - it's not possible to make
the actual blackness visible.

<script>
var t = async_test("Initial state is transparent black");
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

_assertPixel(offscreenCanvas, 20,20, 0,0,0,0, "20,20", "0,0,0,0");

t.done();

});
</script>
